Inline list reduce lambda binding
